Added signout confirmation

+++ snap_n_claim/lib/screens/finance_admin/finance_admin_menu_drawer.dart
@@ -73,14 +73,35 @@ class FinanceAdminMenuDrawer extends StatelessWidget {
   void _onTapMyClaimsBtn(BuildContext context) {}
 
   Future<void> _onTapSignOutBtn(BuildContext context) async {
-    final directory = await getApplicationDocumentsDirectory();
-    final path = directory.path;
-    File file = File('$path/userdata.txt');
-    await file.delete();
+    var dialogRes = await showDialog<bool>(
+        context: context,
+        builder: (context) => AlertDialog(
+              actionsAlignment: MainAxisAlignment.spaceBetween,
+              content: const Text('Are you sure you want to signout?'),
+              actions: <Widget>[
+                TextButton(
+                  onPressed: () => Navigator.pop(context, false),
+                  child: const Text('No'),
+                ),
+                TextButton(
+                  onPressed: () async {
+                    return Navigator.pop(context, true);
+                  },
+                  child: const Text('Yes'),
+                ),
+              ],
+            ));
 
-    Navigator.of(context).pop();
-    Navigator.of(context).pushReplacement(MaterialPageRoute(
-        builder: (BuildContext context) => LoginScreen(_width, _height)));
+    if (dialogRes == true) {
+      final directory = await getApplicationDocumentsDirectory();
+      final path = directory.path;
+      File file = File('$path/userdata.txt');
+      await file.delete();
+
+      Navigator.of(context).pop();
+      Navigator.of(context).pushReplacement(MaterialPageRoute(
+          builder: (BuildContext context) => LoginScreen(_width, _height)));
+    }
   }
 
   @override
@@ -167,7 +188,7 @@ class FinanceAdminMenuDrawer extends StatelessWidget {
         Padding(
           padding: EdgeInsets.symmetric(
             horizontal: _width / 19.63636363636364,
-            vertical: 50,
+            vertical: _height / 16.05818181818182,
           ), // 20
           child: ElevatedButton(
             onPressed: () {
